Update CAN Stack to latest AgIsoStack
Updated AgIsoStack to 29dab887a48bb204aae983b06052d52b0f2314d5

+++ src/isobus_device_descriptor_object_pool.hpp
@@ -4,7 +4,7 @@
 /// @brief Defines an interface for creating a Task Controller DDOP.
 /// @author Adrian Del Grosso
 ///
-/// @copyright 2023 Adrian Del Grosso
+/// @copyright 2023 The Open-Agriculture Developers
 //================================================================================================
 
 #ifndef ISOBUS_DEVICE_DESCRIPTOR_OBJECT_POOL_HPP
@@ -174,9 +174,10 @@ namespace isobus
 		/// @brief Clears the DDOP back to an empty state
 		void clear();
 
-		/// @brief Returns the number of objects in the DDOP
+		/// @brief Returns the number of objects in the DDOP.
+		/// @note The number of objects in the DDOP is limited to 65535.
 		/// @returns The number of objects in the DDOP
-		std::size_t size() const;
+		std::uint16_t size() const;
 
 	private:
 		/// @brief Checks to see that all parent object IDs correspond to an object in this DDOP
